pyproject.toml
[project]
name = "keel"
version = "0.1.0"
description = "Compress the research-to-draft pipeline for long-form analytical writing"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "feedparser>=6.0",
    "httpx>=0.27",
    "pymupdf>=1.24",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
keel = "keel.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/keel"]
